+++ internal/management/controller/roles/postgres.go
@@ -27,6 +27,7 @@ import (
 	"github.com/lib/pq"
 
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
+	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/postgres/utils"
 )
 
 // PostgresRoleManager is a RoleManager for a database instance
@@ -125,7 +126,7 @@ func (sm PostgresRoleManager) Update(ctx context.Context, role DatabaseRole) err
 	// will change no matter what, the next reconciliation cycle we would update the password
 	appendPasswordOption(role, &query)
 
-	if err := executeWithLocalCommit(sm.superUserDB, query.String()); err != nil {
+	if err := utils.ExecuteStatementWithLocalCommit(sm.superUserDB, query.String()); err != nil {
 		return wrapErr(err)
 	}
 	return nil
@@ -150,7 +151,7 @@ func (sm PostgresRoleManager) Create(ctx context.Context, role DatabaseRole) err
 	// NOTE: defensively we might think of doing CREATE ... IF EXISTS
 	// but at least during development, we want to catch the error
 	// Even after, this may be "the kubernetes way"
-	if err := executeWithLocalCommit(sm.superUserDB, query.String()); err != nil {
+	if err := utils.ExecuteStatementWithLocalCommit(sm.superUserDB, query.String()); err != nil {
 		return wrapErr(err)
 	}
 
@@ -159,7 +160,7 @@ func (sm PostgresRoleManager) Create(ctx context.Context, role DatabaseRole) err
 		query.WriteString(fmt.Sprintf("COMMENT ON ROLE %s IS %s",
 			pgx.Identifier{role.Name}.Sanitize(), pq.QuoteLiteral(role.Comment)))
 
-		if err := executeWithLocalCommit(sm.superUserDB, query.String()); err != nil {
+		if err := utils.ExecuteStatementWithLocalCommit(sm.superUserDB, query.String()); err != nil {
 			return wrapErr(err)
 		}
 	}
@@ -177,7 +178,7 @@ func (sm PostgresRoleManager) Delete(ctx context.Context, role DatabaseRole) err
 
 	query := fmt.Sprintf("DROP ROLE %s", pgx.Identifier{role.Name}.Sanitize())
 	contextLog.Debug("Dropping", "query", query)
-	if err := executeWithLocalCommit(sm.superUserDB, query); err != nil {
+	if err := utils.ExecuteStatementWithLocalCommit(sm.superUserDB, query); err != nil {
 		return wrapErr(err)
 	}
 
@@ -218,7 +219,7 @@ func (sm PostgresRoleManager) UpdateComment(ctx context.Context, role DatabaseRo
 	query := fmt.Sprintf("COMMENT ON ROLE %s IS %s",
 		pgx.Identifier{role.Name}.Sanitize(), pq.QuoteLiteral(role.Comment))
 	contextLog.Debug("Updating comment", "query", query)
-	if err := executeWithLocalCommit(sm.superUserDB, query); err != nil {
+	if err := utils.ExecuteStatementWithLocalCommit(sm.superUserDB, query); err != nil {
 		return wrapErr(err)
 	}
 
@@ -259,32 +260,17 @@ func (sm PostgresRoleManager) UpdateMembership(
 		)
 	}
 
-	tx, err := sm.superUserDB.BeginTx(ctx, nil)
-	if err != nil {
-		return wrapErr(err)
-	}
-	defer func() {
-		rollbackErr := tx.Rollback()
-		if rollbackErr != nil && rollbackErr != sql.ErrTxDone {
-			contextLog.Error(rollbackErr, "rolling back transaction")
+	return utils.ExecuteWithLocalCommit(sm.superUserDB, func(tx *sql.Tx) error {
+		for _, sqlQuery := range queries {
+			contextLog.Debug("Executing query", "sqlQuery", sqlQuery)
+			if _, err := sm.superUserDB.ExecContext(ctx, sqlQuery); err != nil {
+				contextLog.Error(err, "executing query", "sqlQuery", sqlQuery, "err", err)
+				return wrapErr(err)
+			}
 		}
-	}()
 
-	// we don't want to be stuck here if synchronous replicas are still not alive
-	// and kicking
-	_, err = tx.Exec("SET LOCAL synchronous_commit to LOCAL")
-	if err != nil {
-		return err
-	}
-
-	for _, sqlQuery := range queries {
-		contextLog.Debug("Executing query", "sqlQuery", sqlQuery)
-		if _, err := sm.superUserDB.ExecContext(ctx, sqlQuery); err != nil {
-			contextLog.Error(err, "executing query", "sqlQuery", sqlQuery, "err", err)
-			return wrapErr(err)
-		}
-	}
-	return tx.Commit()
+		return nil
+	})
 }
 
 // GetParentRoles get the in roles of this role
@@ -315,32 +301,6 @@ func (sm PostgresRoleManager) GetParentRoles(
 	}
 
 	return parentRoles, nil
-}
-
-// executeWithLocalCommit execute the query with commit ack only when WAL flush to local
-// synchronous_commit=local
-func executeWithLocalCommit(db *sql.DB, statement string) error {
-	tx, err := db.Begin()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		// This has no effect if the transaction
-		// is committed
-		_ = tx.Rollback()
-	}()
-
-	// we don't want to be stuck here if synchronous replicas are still not alive
-	// and kicking
-	if _, err := tx.Exec("SET LOCAL synchronous_commit to LOCAL"); err != nil {
-		return err
-	}
-
-	if _, err := tx.Exec(statement); err != nil {
-		return err
-	}
-
-	return tx.Commit()
 }
 
 func appendInRoleOptions(role DatabaseRole, query *strings.Builder) {
